Agent team audit: 11 bugs fixed, 7 feature alignments, 611 tests passing (batch 21)
Code audit (4 bugs):
- Fix benchmark.py: fetch_stock_data -> fetch_historical with date formatting
- Fix data_fetcher.py: yfinance FastInfo attribute access (not dict .get())
- Fix feature_engineer.py: bool.astype() -> int() for earnings proximity
- Fix metrics.py: monthly return column mislabeling with positional slicing

Feature alignment (7 locations):
- backtester.py: 3 methods had sentiment/macro/interactions=True
- degradation_monitor.py: fresh data collection had all flags True
- walk_forward_optimizer.py: 3 locations with wrong flags + Optuna search space
- train_mlp.py: migrated from add_all_features() to add_all_features_extended()

Test suite (7 files fixed, 83 failures -> 0):
- Rewrote test_broker, test_strategy, test_models, test_regime_detector,
  test_data_fetcher to match current API
- Fixed conftest fixtures, test_feature_engineer, test_analytics assertions

+++ src/data/feature_engineer.py
@@ -768,7 +768,7 @@ class FeatureEngineer:
             # Days to earnings
             if fundamentals.days_to_earnings is not None:
                 df["days_to_earnings"] = float(fundamentals.days_to_earnings)
-                df["near_earnings"] = (fundamentals.days_to_earnings <= 14).astype(int)
+                df["near_earnings"] = int(fundamentals.days_to_earnings <= 14)
             else:
                 df["days_to_earnings"] = 90.0
                 df["near_earnings"] = 0
